+++ test/fixtures/reportCliError/CliError-instance.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const CliError = require('../../../private/CliError');
+const reportCliError = require('../../../private/reportCliError');
+
+reportCliError('CLI', new CliError('Message.'));