+++ test/fixtures/reportCliError/Error-instance-without-stack.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const reportCliError = require('../../../private/reportCliError');
+
+const error = new Error('Message.');
+delete error.stack;
+reportCliError('CLI', error);