+++ test/fixtures/reportCliError/primitive-value.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const reportCliError = require('../../../private/reportCliError');
+
+reportCliError('CLI', '');