+++ test/index.js
@@ -8,6 +8,7 @@ require('./replaceStackTraces.test')(tests);
 require('./cli/find-unused-exports.test')(tests);
 require('./private/getVariableDeclarationIdentifierNames.test')(tests);
 require('./private/isDirectoryPath.test')(tests);
+require('./private/reportCliError.test')(tests);
 require('./private/scanModuleCode.test')(tests);
 require('./private/scanModuleFile.test')(tests);
 require('./public/findUnusedExports.test')(tests);